serverless-endpoint/CLIP-ViT-B-32/code/requirements.txt
sentence-transformers==2.2.2
Pillow==10.0.0
torch==1.13.1
torchvision==0.14.1